requirements.txt
djangorestframework == 3.15.2
djangorestframework-simplejwt == 5.3.1
celery == 5.4.0
redis==5.0.7
django-celery-results == 2.5.1
django-celery-beat == 2.6.0
python-dotenv==1.0.1
django-redis==5.4.0